+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.19](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.18...@ovh-ux/manager-pci-app@3.7.19) (2023-02-23)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [3.7.18](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.17...@ovh-ux/manager-pci-app@3.7.18) (2023-02-20)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app